third_party/blink/web_tests/tables/hittesting/filltable-emptycells-expected.txt

SUCCESS!!!
Hit cell at row: 0, column: 0: SUCCESS
Hit cell at row: 0, column: 1: EMPTY
Hit cell at row: 0, column: 2: EMPTY
Hit cell at row: 1, column: 0: SUCCESS
Hit cell at row: 1, column: 1: SUCCESS
Hit cell at row: 1, column: 2: SUCCESS
